yolov8/requirements.txt
ipywidgets
torch>=2.1
torchvision>=0.16
ultralytics==8.0.43
openvino
onnxruntime==1.16.3
onnxruntime-openvino==1.16